Separate librustcdoc module
